clk: mvebu: add AP806 core clock driver
This commit adds a new driver to handle the core clocks found in the
AP806 HW block, which is the core block of all Armada 7K and 8K
Marvell 64-bits processors. This core clock driver reads the
Sample-At-Reset register to determine the frequencies of several core
clocks: DDR, Ring, CPU, PIDI AP and PIDI CP clocks.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>