Revert "ARM: mm: fix support for HW coherent systems in PL310 cache"
This patch reverts the PL310 removal of L2 operations due to HWCC
system. This is done because of a new (and dymanic) way to find this
configuration and remove the L2 operations.
The new implementation is introduced in the followin patches.

This reverts 'commit c4e4d88303a0 ("ARM: dma-mapping: Don't use
outer_flush_range when the L2C is coherent")'.

Change-Id: I4bd2c87396971a7f0969441f2d7c1b42d010bb09
Signed-off-by: Lior Amsalem <alior@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/28513
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Ofer Heifetz <oferh@marvell.com>