irqchip: irq-mvebu-advk: add new driver
Armada 3700 SoC's PCIe controller comprises an internal IRQ controller
capable of routing both legacy INTX and MSI interrupts to the one of
GIC SPI's.

This driver adds legacy interrupt support in a new driver, as well as
an according documentation.

The code is based on initial work of Hezi Shahmoon.

Change-Id: I74f5c24169674b841c192ae728c26d2d26479e08
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Signed-off-by: Ofer Heifetz <oferh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/28766
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Nadav Haklai <nadavh@marvell.com>