net: mvpp2x: silence complaints on probe deferral
On some platforms, the clocks might be registered by a platform
driver. When this is the case, the clock platform driver may very well
be probed after mvpp2 driver, in which case the first probe() invocation
of mvpp2 driver will receive -EPROBE_DEFER as the return value of
devm_clk_get().

The current code handles that as a normal error, and prints error message.
This patch silence the error message, in case of probe deferral.

This patch is part of patch-set which adds
mainline mvebu clock gating support [14/21].

Change-Id: Iea94b1e7d542ed50f6d66c88be79f7c5a6351047
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32399
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>