arm64: dts: marvell: add pinctrl description for rgmii
Placing pinctrl description, inside one of the child node of 'ppv22' would
be more clear. Unfortunately pinctrl description can not be moved to the
child nodes of 'ppv22', because the pinctrl framework requires that the
description is placed inside the device nodes (it is because all pinctrl
routines are called during driver_probe_device). In case of ppv2, there is
only one driver registered for whole ppv22, which is common for each port.

Change-Id: Ia94bd7cd7438cfe3df2029d5bd17cb6af7ea71cc
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32934
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>